Fix for #945, cpu temperature is signed.
Added a type conversion to cpu temperature sysctl.  Will still
collect/report -1 when the value is -1, this is because it should be up
to interpretation whether this is the correct value for the system or
not.

Some drivers will report -1 for cpu temperature.  Other sensors will
report "an input into the fan control algorithm", i.e. not the actual
temperature, but how much fan it wants.  Some people cool their machines
with liquid nitrogen.

+++ collector/cpu_freebsd.go
@@ -138,7 +138,8 @@ func (c *statCollector) Update(ch chan<- prometheus.Metric) error {
 			}
 			continue
 		}
-		ch <- c.temp.mustNewConstMetric(float64(temp-2732)/10, lcpu)
+		// Sysctl temp is an signed integer, so we do the conversion.
+		ch <- c.temp.mustNewConstMetric(float64(int32(temp)-2732)/10, lcpu)
 	}
 	return err
 }
